style: extract HealthCheck to core/health.py, strip overlay docstrings
- Extract HealthCheck, default_health_checks to core/health.py (no lint suppression needed)
- Strip inline docstrings from overlay.py (names are the docs per ac-python)
- overlay.py: 603 → 339 LOC
